remove unnecessary middleware
